Improve CRUD code after Marcus's feedback
