+++ portfolio/src/main/webapp/index.html
@@ -30,16 +30,16 @@
     <body onload="loadTasks()">
         <h1>Todo List</h1>
     
-        <form method="POST" action="/new-task">
+        <form method="POST" action="/tasks">
           <input type="text" name="title" />
           <br/>
-          <button>Submit</button>
+          <button onclick="">Submit</button>
         </form>
     
         <hr/>
     
         <ul id="task-list"></ul>
     
-      </body>
+    </body>
   </body>
 </html>
